linear algebra parallelization made independend of its parent communicator by introducing ortho_parent_comm to be used when addressing the whole group. linear algebra is now distributed (in PW) inside the pool group (in CPV is left unchanged... are there pools in CPV?). mp_global sets ortho_comm as a sub-communicator of intra_pool_comm (used to be intra_bgrp_comm). It can be reverted to previous choice by commenting/uncommenting one line
tested on PW/example02 co.rx.in case (both Gamma and K=(000)) with 
  -np 8 -nd 4 -nb 2
that is using 2 bgrp (procs 0123 and 4567) and diagonalizing on 4 procs (0246).
tested also on 
  -np 4 -nd 4 -nb 2
that is using 2 bgrp (procs 01 and 34) and diagonalizing on 4 procs (0123).

some bgrp parallelization added to a few routines. global variables (evc,..) are NOT distributed but some local ones
are and more could be done.




git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@11863 c92efa57-630b-4861-b058-cf58834340f0