temporary fix for test case pw_berry.
In realus the l-dependednt q(r) is built (now actually copied from qfuncl), all works fine
and in the last stage of the routine the q(r) on the radial grid is used to fill the relevant 
FFT grid pointsi for each atom. 

This is done by a spline interpolation iexcept for the original vandebilt case (with rinner > 0)
where one can also use directly the qfcoef expansion to compute the value.

It shouldn't matter but in the pw_berry test case this change the energy by 1.4d-5, enough to make it fail.

Added a flag that keeps using the original method until a better understanding of the issue is reached.



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@12601 c92efa57-630b-4861-b058-cf58834340f0