Noncollinear magnetisation support with USPP and PAW pseudopotentials implemented. This version was developed by Takashi Koretsune and Florian Thoele, another version by X Zubizarreta and M Calandra was used for testing.
NOTE: nnkp files produced by older versions (<2.0) of wannier90.x can still be used, but the workflow is a bit tricky and we recommend rerunning wannier90 instead.
Compatibility will be dropped soon, as it adds a lot of complexity and makes the code behaviour quite difficult to predict.

The code required to ensure the compatibility is only executed whn the old_spinor_proj flag is set to true. 



git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/q-e/trunk/espresso@12997 c92efa57-630b-4861-b058-cf58834340f0